requirements.txt
# Core Framework
dash==2.16.1
dash-bootstrap-components==1.5.0
plotly==5.17.0
flask==2.3.3
flask-cors==4.0.0
flask-compress==1.13
flask-limiter==3.5.0
gunicorn==21.2.0
gevent==23.7.0

# Database
pymongo==4.5.0
redis==5.0.0

# Security & Authentication
bcrypt==4.1.1
pyjwt==2.8.0
pyotp==2.9.0
qrcode==7.4.2
cryptography==41.0.7
werkzeug==2.3.7

# AI/ML (optional - comment out if not needed)
# tensorflow==2.15.0
# torch==2.1.1
# transformers==4.36.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# Document Processing
reportlab==4.0.7
Pillow==10.1.0

# Monitoring
psutil==5.9.6

# Production Utilities (optional)
celery==5.3.4
python-dotenv==1.0.0 
